Split restart-stable cells by responsibility
`cells.rs` had grown to 3630 lines carrying four unrelated
responsibilities behind one module boundary: the observation gate and its
transaction capability, the retained-exit guards, the member cell, and the
scope cell with its observation projection.

Split it into `cells/{mod,gate,member,retained,scope}.rs` plus
`cells/scope/projection.rs`. The move is mechanical — every item keeps its
text, its declaration order and its struct field order, which the
`ObservationTxn` guard/effects order and the raw-value-before-guards order
in the retained records both depend on.

`mod.rs` keeps the module header, the four submodule declarations, the glob
re-exports that hold `lib.rs`'s explicit export list unchanged, and a
`test_support` module for the fixtures several child test modules share.

Visibility was re-derived from private upward: the 26 widenings are all
private -> `pub(super)`, and the four in `scope/projection.rs` are narrower
than the module-private scope they replace. Nothing became `pub` or
`pub(crate)`.

AGENTS.md's lock rule now points at `cells/gate.rs`, where `ObservationTxn`
lives.
